Update, testing and planing
Added a reminder slot to in the excel and for the user to input when creating a event.

Next up:
Want: Implement a function check_events if todays date is the day the user want's to get reminded of a specific event.
Problem: We have to convert the information (day, month, year) from the excel Database to a date object.

I did this at the bottom of the code. Next is to implement the calculation for the reminder function and integrate the conversion into the function.